feat: detect firmware vendor transitions within a BBL file
When a BBL file contains sessions from multiple firmware vendors (e.g.
Betaflight sessions left on flash before reflashing to EmuFlight),
bbl_parser now detects the transition and:

- Prints a WARNING listing session ranges per vendor with the suggestion
  to erase flash before reflashing
- Corrects the output filename prefix for sessions whose firmware vendor
  does not match the BBL filename prefix (e.g. EMUF_* file + Betaflight
  session → output named BTFL_*), preventing EMUF_BTFL_* or BTFL_BTFL_*
  duplicate prefixes

New public API in export module:
- `firmware_prefix_for_revision(revision)` — maps a firmware revision
  string to its canonical filename prefix (EMUF_, BTFL_, INAV_)
- `corrected_session_base_name(bbl_path, firmware_revision)` — returns
  a corrected base name when the session vendor differs from the BBL
  filename prefix, or None when they match

All export functions (`export_to_csv`, `export_to_gpx`, `export_to_event`,
`compute_export_paths`) accept an optional `base_name_override` parameter.
Existing callers pass `None` to preserve prior behaviour.

Known firmware prefixes are defined in one place (`KNOWN_FIRMWARE_PREFIXES`)
and extended by adding a single `(&str, &str)` entry.

+++ src/conversion.rs
@@ -343,8 +343,7 @@ fn parse_datetime_to_epoch(datetime_str: &str) -> Option<u64> {
 
     // Convert to days since epoch (simplified, doesn't handle all edge cases)
     let days = ymd_to_days(year, month, day)?;
-    let local_secs =
-        (days as u64) * 86400 + (hour as u64) * 3600 + (minute as u64) * 60 + (second as u64);
+    let local_secs = days * 86400 + (hour as u64) * 3600 + (minute as u64) * 60 + (second as u64);
 
     // Convert local time to UTC by subtracting the offset
     // If offset is +02:00, local time is 2 hours ahead of UTC, so subtract 2 hours
